move functions to different modules and call com1DFA also from runscript for Oscars branch particle analysis plots
